discord-github-worker/requirements.txt
discord.py>=2.0.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
